# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API бронирования",
        "description": "Документация API бронирования ресурсов",
        "version": "1.0"
    },
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Создать бронирование",
                "operationId": "6e1e75ede753a8655c505cd98c57c7ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resource_id",
                                    "user_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "resource_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-21T14:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-02-21T16:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Бронирование создано"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Отменить бронирование",
                "operationId": "a89472407dc26914b54cf8f9751aa43c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование отменено"
                    }
                }
            }
        },
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получить список ресурсов",
                "operationId": "dcd8a548291f6975f165697969797dae",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Создать ресурс",
                "operationId": "de8c8265c94763fd4ee1ee535b7d0f66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Велосипед"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "bike"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Велосипед 21 скорость"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан"
                    }
                }
            }
        },
        "/api/resources/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получить ресурс по ID",
                "operationId": "d3788b011b3b541ed1fc68fbd49178d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ресурса",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Управление бронированиями"
        },
        {
            "name": "Resources",
            "description": "Управление ресурсами"
        }
    ]
}